feat : pause and option menu 🔢  fix : canvas layers 👍
ESC key to pause, options menu coming next.

You can now access the Item tooltips while dialogue is active. This was accomplished by lower the dialogue balloon canvas layer. Because the balloon canvas layer was covering the entire screen, tooltips did not sense the mouse had entered the item sprite. I was in the process of troubleshooting the buttons of the pause menu when I noticed Balloon has a canvas layer.